Set the maximum users in an org
Set the maximum number of users an org can have with the new max_users property on the Org object. Set it with the usual updateOrg() function, and see the current value with fetchOrg(). It's an optional field, and if it's missing orgs will continue to allow unlimited users.

Let only certain users create orgs

Users can be given permission to create orgs through enableUserCanCreateOrgs, and that permission can be removed with disableUserCanCreateOrgs. This special permission overrides the realm-wide users_can_create_orgs. With this new functionality, you can turn off org creation in general, and only grant certain users the privilege. (If the realm-wide users_can_create_orgs is on, then everyone will always be able to create orgs.)

API Keys

A whole host of functions for managing API Keys. These aren't the Integration API Keys you use to communicate with PropelAuth (in this very library), these are API Keys that you can give out to your own clients. We'll happily manage them for you. You can set the optional user, an optional org, optional metadata, and an expiration.

These new API Key endpoints are:

- fetchApiKey
- fetchCurrentApiKeys
- fetchArchivedApiKeys
- createApiKey
- updateApiKey
- deleteApiKey
- validateApiKey